PCI: Convert dynamic "rom" sysfs object into static
The "rom" sysfs object, which is a read-only binary attribute, resides
at the "/sys/bus/pci/devices/.../rom" path and allows for accessing the
PCI ROM if the device has one.

At the moment, the "rom" sysfs object is dynamically created when the
pci_create_sysfs_dev_files() function executes that is part of the PCI
sysfs objects initialisation and when a device is added:

  late_initcall()
    pci_sysfs_init()
      pci_create_sysfs_dev_files()
        sysfs_bin_attr_init()
        sysfs_create_bin_file()

  pci_bus_add_devices()
    pci_bus_add_device()
      pci_create_sysfs_dev_files()
        ...

And dynamically removed when the pci_remove_sysfs_dev_files() function
executes when the PCI device is stopped and removed, as per:

  pci_stop_bus_device()
    pci_stop_dev()
      pci_remove_sysfs_dev_files()
        sysfs_remove_bin_file()

This attribute does not need to be created dynamically and removed
dynamically, and thus there is no need to also manage its create and
remove life cycle manually as the PCI driver core can do it when the
device is either added or removed.

Convert the "rom" sysfs object created dynamically to static and use the
.is_bin_visible callback to set the correct sysfs object size based on
the PCI ROM size.

Then, add a newly created attribute group to the existing group called
"pci_dev_groups" to ensure that the "rom" sysfs object would be
automatically included when the PCI driver initializes.

Remove the no longer needed pointer "rom_attr" from the struct pci_dev
that used to hold a reference to the dynamically created "rom" sysfs
object.

Suggested-by: Oliver O'Halloran <oohall@gmail.com>
Suggested-by: Bjorn Helgaas <bhelgaas@google.com>
Signed-off-by: Krzysztof Wilczyński <kw@linux.com>