alpha/PCI: Lock down BAR access when the kernel is locked down
Any hardware that can potentially generate DMA has to be locked down in
order to avoid it being possible for an attacker to modify kernel code,
allowing them to circumvent disabled module loading or module signing.
Default to paranoid - in future we can potentially relax this for
sufficiently IOMMU-isolated devices.

Related:
  commit eb627e17727e ("PCI: Lock down BAR access when the kernel is locked down")

Signed-off-by: Krzysztof Wilczyński <kw@linux.com>