PCI/sysfs: Rename pci_read_resource_io() and pci_write_resource_io()
The pci_read_resource_io() and pci_write_resource_io() functions are
used for the corresponding read() and write() callbacks when declaring
and defining a static sysfs object for a given PCI resource.

Currently, only reading and writing against an I/O BARs is supported,
but we might change this in the future to also include memory-mapped
resources.

Thus, rename to read() and write() callbacks to be more generic and drop
the I/O BARs-specific suffix from them.

Signed-off-by: Krzysztof Wilczyński <kw@linux.com>